#11 Split server.go
