+++ backend/models/add_item_request.go
@@ -1,0 +1,6 @@
+package models
+
+type AddItemRequest struct {
+	ItemId string `json:"itemId"`
+	Amount int    `json:"amount"`
+}
