+++ backend/models/locales.go
@@ -1,0 +1,5 @@
+package models
+
+type Locales struct {
+	Data map[string]string `json:"data"`
+}
