+++ backend/models/server_info.go
@@ -1,0 +1,7 @@
+package models
+
+type ServerInfo struct {
+	Version    string `json:"version"`
+	Path       string `json:"path"`
+	ModVersion string `json:"modVersion"`
+}
