+++ backend/spt/server.go
@@ -1,165 +1,66 @@
 package spt
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
 	"slices"
 	"sort"
+	"spt-give-ui/backend/http"
+	"spt-give-ui/backend/models"
+	"spt-give-ui/backend/util"
 	"strings"
-	"time"
 )
 
-type ServerInfo struct {
-	Version    string `json:"version"`
-	Path       string `json:"path"`
-	ModVersion string `json:"modVersion"`
-}
-
-type SPTProfileInfo struct {
-	Id       string `json:"id"`
-	Username string `json:"username"`
-}
-
-type SPTProfile struct {
-	Info SPTProfileInfo `json:"info"`
-}
-
-type BSGItem struct {
-	Id     string       `json:"_id"`
-	Parent string       `json:"_parent"`
-	Type   string       `json:"_type"`
-	Props  BSGItemProps `json:"_props"`
-}
-
-type BSGItemProps struct {
-	StackMaxSize int  `json:"StackMaxSize"`
-	IsUnbuyable  bool `json:"IsUnbuyable"`
-}
-
-type Locales struct {
-	Data map[string]string `json:"data"`
-}
-
-type ViewItem struct {
-	Id          string
-	Name        string
-	Type        string
-	Description string
-	Category    string
-	MaxStock    int
-}
-
-type AllItems struct {
-	Categories []string
-	Items      []ViewItem
-	Presets    []ViewItem
-}
-
-type AddItemRequest struct {
-	ItemId string `json:"itemId"`
-	Amount int    `json:"amount"`
-}
-
-var myClient = &http.Client{Timeout: 10 * time.Second}
-
-func doPost(url string, sessionId string, body interface{}) (*http.Response, error) {
-	jsonBody, err := json.Marshal(body)
-	if err != nil {
-		return nil, err
-	}
-	req, err := http.NewRequest("POST", url, bytes.NewReader(jsonBody))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("responsecompressed", "0")
-	req.Header.Set("requestcompressed", "0")
-	req.Header.Set("Cookie", fmt.Sprintf("PHPSESSID=%s", sessionId))
-	return myClient.Do(req)
-}
-
-func doGet(url string) (*http.Response, error) {
-	req, _ := http.NewRequest("GET", url, nil)
-	req.Header.Set("responsecompressed", "0")
-	return myClient.Do(req)
-}
-
-func getJson(url string, target interface{}) error {
-	r, err := doGet(url)
-	if err != nil {
-		return err
-	}
-	defer r.Body.Close()
-	return json.NewDecoder(r.Body).Decode(target)
-}
-
-func getRawBytes(url string) ([]byte, error) {
-	r, err := doGet(url)
-	if err != nil {
-		return nil, err
-	}
-	// Read the entire response body into a byte slice
-	data, err := io.ReadAll(r.Body)
-	if err != nil {
-		return nil, err
-	}
-	return data, nil
-}
-
-func ConnectToSptServer(host string, port string) (r *ServerInfo, e error) {
-	serverInfo := &ServerInfo{}
-	err := getJson(fmt.Sprintf("http://%s:%s/give-ui/server", host, port), serverInfo)
+func ConnectToSptServer(host string, port string) (r *models.ServerInfo, e error) {
+	serverInfo := &models.ServerInfo{}
+	err := util.GetJson(fmt.Sprintf("http://%s:%s/give-ui/server", host, port), serverInfo)
 	if err != nil {
 		return nil, err
 	}
 	return serverInfo, nil
 }
 
-func LoadProfiles(host string, port string) (r []SPTProfileInfo, e error) {
-	profiles, err := getRawBytes(fmt.Sprintf("http://%s:%s/give-ui/profiles", host, port))
+func LoadProfiles(host string, port string) (r []models.SPTProfileInfo, e error) {
+	profiles, err := util.GetRawBytes(fmt.Sprintf("http://%s:%s/give-ui/profiles", host, port))
 	if err != nil {
 		return nil, err
 	}
-	var sessionsMap map[string]SPTProfile
-	err = parseByteResponse(profiles, &sessionsMap)
+	var sessionsMap map[string]models.SPTProfile
+	err = util.ParseByteResponse(profiles, &sessionsMap)
 	if err != nil {
 		return nil, err
 	}
-	var sessions []SPTProfileInfo
+	var sessions []models.SPTProfileInfo
 	for _, v := range sessionsMap {
 		sessions = append(sessions, v.Info)
 	}
 	return sessions, nil
 }
 
-func LoadItems(host string, port string) (r *AllItems, e error) {
-	itemsBytes, err := getRawBytes(fmt.Sprintf("http://%s:%s/give-ui/items", host, port))
+func LoadItems(host string, port string) (r *models.AllItems, e error) {
+	itemsBytes, err := util.GetRawBytes(fmt.Sprintf("http://%s:%s/give-ui/items", host, port))
 	if err != nil {
 		return nil, err
 	}
-	var itemsMap map[string]BSGItem
-	err = parseByteResponse(itemsBytes, &itemsMap)
+	var itemsMap map[string]models.BSGItem
+	err = util.ParseByteResponse(itemsBytes, &itemsMap)
 	if err != nil {
 		return nil, err
 	}
 	// TODO hardcoded locale en
-	localeBytes, err := getRawBytes(fmt.Sprintf("http://%s:%s/client/locale/en", host, port))
+	localeBytes, err := util.GetRawBytes(fmt.Sprintf("http://%s:%s/client/locale/en", host, port))
 	if err != nil {
 		return nil, err
 	}
-	var locales Locales
-	err = parseByteResponse(localeBytes, &locales)
+	var locales models.Locales
+	err = util.ParseByteResponse(localeBytes, &locales)
 	if err != nil {
 		return nil, err
 	}
 
-	allItems := AllItems{
+	allItems := models.AllItems{
 		Categories: []string{},
-		Items:      []ViewItem{},
-		Presets:    []ViewItem{},
+		Items:      []models.ViewItem{},
+		Presets:    []models.ViewItem{},
 	}
 	for _, bsgItem := range itemsMap {
 		if bsgItem.Type == "Node" || bsgItem.Props.IsUnbuyable {
@@ -190,7 +91,7 @@ func LoadItems(host string, port string) (r *AllItems, e error) {
 			continue
 		}
 
-		viewItem := ViewItem{
+		viewItem := models.ViewItem{
 			Id:          bsgItem.Id,
 			Name:        name,
 			Type:        bsgItem.Type,
@@ -212,17 +113,9 @@ func LoadItems(host string, port string) (r *AllItems, e error) {
 }
 
 func AddItem(host string, port string, sessionId string, itemId string, amount int) {
-	request := AddItemRequest{
+	request := models.AddItemRequest{
 		ItemId: itemId,
 		Amount: amount,
 	}
-	doPost(fmt.Sprintf("http://%s:%s/give-ui/give", host, port), sessionId, request)
-}
-
-func parseByteResponse(profiles []byte, target interface{}) error {
-	err := json.Unmarshal(profiles, target)
-	if err != nil {
-		return err
-	}
-	return nil
+	http.DoPost(fmt.Sprintf("http://%s:%s/give-ui/give", host, port), sessionId, request)
 }
